Add more default excludes and analysis languages
This commit adds a list of default excluded files,
including most well-known binary files. All of these
files will be excluded from analysis.

Also, add a new parameter to the directory class
that allows users to specify a list of anaysis
languages. Any file that is not one of these
analysis languages will be ignored.